+++ packages/editor/editor-core/src/labs/next/internal/components/EditorInternal.tsx
@@ -3,7 +3,7 @@ import { intlShape } from 'react-intl';
 import * as PropTypes from 'prop-types';
 import { WidthProvider } from '@atlaskit/editor-common';
 import EditorContext from '../../../../ui/EditorContext';
-import { EditorActions } from '../../../../index';
+import EditorActions from '../../../../actions';
 import { PortalProviderAPI } from '../../../../ui/PortalProvider';
 import { EditorProps } from '../editor-props-type';
 import { EditorSharedConfigProvider } from '../context/shared-config';
@@ -11,15 +11,25 @@ import { useEditor } from '../hooks/use-editor';
 import { EditorContentProvider } from './EditorContent';
 
 export function EditorInternal(props: EditorPropsExtended, context: any) {
-  const editorActions =
-    ((context || {}).editorActions as EditorActions) || new EditorActions();
+  // Need to memoize editor actions otherwise in case when editor is not
+  // wrapped with EditorContext every prop change triggers all hooks
+  // that depend on editorActions
+  const maybeEditorActions = (context || {}).editorActions;
+  const editorActions = React.useMemo(
+    () => maybeEditorActions || new EditorActions(),
+    [maybeEditorActions],
+  );
 
   const [editorSharedConfig, mountEditor] = useEditor({
     context,
     editorActions,
+    handleAnalyticsEvent: props.handleAnalyticsEvent,
 
     disabled: props.disabled,
+
+    transformer: props.transformer,
     defaultValue: props.defaultValue,
+
     plugins: props.plugins,
 
     portalProviderAPI: props.portalProviderAPI,
@@ -28,7 +38,10 @@ export function EditorInternal(props: EditorPropsExtended, context: any) {
     popupsScrollableElement: props.popupsScrollableElement,
 
     onChange: props.onChange,
+    onDestroy: props.onDestroy,
   });
+
+  const onMount = props.onMount;
 
   React.useEffect(
     () => {
@@ -37,12 +50,15 @@ export function EditorInternal(props: EditorPropsExtended, context: any) {
           editorSharedConfig.editorView,
           editorSharedConfig.eventDispatcher,
         );
-        return () => {
-          editorActions._privateUnregisterEditor();
-        };
+
+        if (onMount) {
+          onMount(editorActions);
+        }
+
+        return () => editorActions._privateUnregisterEditor();
       }
     },
-    [editorSharedConfig, editorActions],
+    [editorSharedConfig, editorActions, onMount],
   );
 
   return (