+++ packages/editor/editor-core/src/labs/next/internal/hooks/use-editor/create-editor.ts
@@ -1,5 +1,6 @@
 import { EditorView } from 'prosemirror-view';
 import { EditorState } from 'prosemirror-state';
+import { CreateUIAnalyticsEvent } from '@atlaskit/analytics-next';
 import { ProviderFactory } from '@atlaskit/editor-common';
 import {
   EventDispatcher,
@@ -17,6 +18,8 @@ import { EditorProps } from '../../editor-props-type';
 
 export function createEditor({
   context,
+  handleAnalyticsEvent,
+  transformer,
 
   plugins,
   portalProviderAPI,
@@ -29,6 +32,7 @@ export function createEditor({
 
   disabled,
   onChange,
+  onDestroy,
 }: CreateEditorParams): EditorSharedConfig | null {
   if (!ref) {
     return null;
@@ -39,6 +43,7 @@ export function createEditor({
   const dispatch = createDispatch(eventDispatcher);
   const editorConfig = processPluginsList(plugins || [], {});
   const schema = createSchema(editorConfig);
+  const transformerInstance = transformer && transformer(schema);
   const pmPlugins = createPMPlugins({
     editorConfig,
     schema,
@@ -50,13 +55,13 @@ export function createEditor({
     reactContext: () => context,
     dispatchAnalyticsEvent: () => {},
   });
-
   const state = EditorState.create({
     schema,
     plugins: pmPlugins,
-    doc: processRawValue(schema, defaultValue),
+    doc: transformerInstance
+      ? transformerInstance.parse(defaultValue)
+      : processRawValue(schema, defaultValue),
   });
-
   const editorView = new EditorView(
     { mount: ref },
     {
@@ -74,6 +79,9 @@ export function createEditor({
   return {
     editorView,
 
+    transformer: transformerInstance,
+    handleAnalyticsEvent,
+
     eventDispatcher,
     dispatch,
 
@@ -87,6 +95,7 @@ export function createEditor({
     disabled,
     providerFactory,
     onChange,
+    onDestroy,
   };
 }
 
@@ -99,8 +108,12 @@ export type CreateEditorParams = Pick<
   | 'popupsScrollableElement'
   | 'onChange'
   | 'disabled'
+  | 'transformer'
+  | 'handleAnalyticsEvent'
+  | 'onDestroy'
 > & {
   context: any;
   ref?: HTMLDivElement | null;
   portalProviderAPI: PortalProviderAPI;
+  createAnalyticsEvent?: CreateUIAnalyticsEvent;
 };