+++ packages/editor/editor-core/src/plugins/analytics/types/enums.ts
@@ -31,6 +31,7 @@ export enum ACTION {
   BROWSER_FREEZE = 'browserFreeze',
   SLOW_INPUT = 'slowInput',
   UPLOAD_EXTERNAL_FAIL = 'uploadExternalFailed',
+  INPUT_PERF_SAMPLING = 'inputPerfSampling',
 }
 
 export enum INPUT_METHOD {
@@ -70,6 +71,8 @@ export enum ACTION_SUBJECT {
   TEXT = 'text',
   TYPEAHEAD = 'typeAhead',
   TABLE = 'table',
+  EXPAND = 'expand',
+  NESTED_EXPAND = 'nestedExpand',
 }
 
 export enum ACTION_SUBJECT_ID {
@@ -118,6 +121,8 @@ export enum ACTION_SUBJECT_ID {
   STATUS = 'status',
   SYMBOL = 'symbol',
   TABLE = 'table',
+  EXPAND = 'expand',
+  NESTED_EXPAND = 'nestedExpand',
   TYPEAHEAD_EMOJI = 'emojiTypeAhead',
   TYPEAHEAD_LINK = 'linkTypeAhead',
   TYPEAHEAD_MENTION = 'mentionTypeAhead',
@@ -141,4 +146,6 @@ export enum ACTION_SUBJECT_ID {
   PASTE_TABLE_ROW = 'tableRow',
   PASTE_TABLE = 'table',
   PASTE_TASK_LIST = 'taskList',
+  PASTE_EXPAND = 'expand',
+  PASTE_NESTED_EXPAND = 'nestedExpand',
 }