+++ packages/editor/editor-core/src/types/editor-props.ts
@@ -188,6 +188,8 @@ export interface EditorProps {
 
   UNSAFE_cards?: CardOptions;
 
+  UNSAFE_allowExpand?: boolean;
+
   // Submits on the enter key. Probably useful for an inline comment editor use case.
   saveOnEnter?: boolean;
 
@@ -275,4 +277,8 @@ export interface EditorProps {
   // flag to indicate display name instead of nick name should be inserted for mentions
   // default: false, which inserts the nick name
   mentionInsertDisplayName?: boolean;
+
+  // The nth keystroke after which an input time taken event is sent, 0 to disable it
+  // default: 100
+  inputSamplingLimit?: number;
 }