+++ packages/media/media-picker/src/popup/middleware/finalizeUpload.ts
@@ -1,11 +1,9 @@
 import { Store, Dispatch, Middleware } from 'redux';
+import { ReplaySubject } from 'rxjs/ReplaySubject';
 import {
   MediaStore,
   MediaStoreCopyFileWithTokenBody,
   MediaStoreCopyFileWithTokenParams,
-} from '@atlaskit/media-store';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
-import {
   FileState,
   MediaFile as MediaClientFile,
 } from '@atlaskit/media-client';
@@ -15,7 +13,7 @@ import {
 } from '../actions/finalizeUpload';
 import { State, SourceFile } from '../domain';
 import { mapAuthToSourceFileOwner } from '../domain/source-file';
-import { MediaFile } from '../../domain/file';
+import { MediaFile } from '../../types';
 import { sendUploadEvent } from '../actions/sendUploadEvent';
 import { resetView } from '../actions';
 import { UploadEndEvent } from '../../domain/uploadEvent';