+++ packages/media/media-picker/src/service/__tests__/unit/uploadService.spec.ts
@@ -12,21 +12,22 @@ import {
   getFileStreamsCache,
   FileState,
 } from '@atlaskit/media-client';
-import { TouchedFiles } from '@atlaskit/media-client';
-import { AuthProvider, Auth, ProcessingFileState } from '@atlaskit/media-core';
+import { TouchedFiles, ProcessingFileState } from '@atlaskit/media-client';
+import { AuthProvider, Auth } from '@atlaskit/media-core';
 import uuidV4 from 'uuid/v4';
 import { asMock, fakeMediaClient } from '@atlaskit/media-test-helpers';
 import { Observable } from 'rxjs/Observable';
 import { Subscriber } from 'rxjs';
 import { UploadServiceImpl } from '../../uploadServiceImpl';
-import { MediaFile, UploadParams } from '../../..';
 import * as getPreviewModule from '../../../util/getPreviewFromBlob';
 import * as getPreviewFromImage from '../../../util/getPreviewFromImage';
-import { Preview } from '../../../domain/preview';
 import {
+  Preview,
+  MediaFile,
+  UploadParams,
   UploadPreviewUpdateEventPayload,
   UploadsStartEventPayload,
-} from '../../../domain/uploadEvent';
+} from '../../../types';
 
 const fileStreamCacheSpy = jest.spyOn(getFileStreamsCache(), 'set');