+++ packages/editor/editor-wikimarkup-transformer/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @atlaskit/editor-wikimarkup-transformer
 
+## 4.5.8
+
+- Updated dependencies [f28c191f4a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f28c191f4a):
+- Updated dependencies [24b8ea2667](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/24b8ea2667):
+  - @atlaskit/editor-json-transformer@6.3.5
+  - @atlaskit/editor-test-helpers@10.1.3
+  - @atlaskit/editor-core@114.0.0
+  - @atlaskit/renderer@53.0.0
+  - @atlaskit/editor-common@41.2.1
+
+## 4.5.7
+
+### Patch Changes
+
+- [patch][650b9bd18d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/650b9bd18d):
+
+  fix up issue in wikimarkup parser where a url with url-encoded values are double encoded when converted into ADF
+
+## 4.5.6
+
+### Patch Changes
+
+- [patch][24165e116f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/24165e116f):
+
+  fix bug where bullet points and rulers inside monospace causes parse error
+
 ## 4.5.5
 
 ### Patch Changes